ingest/requirements.txt
psycopg2-binary==2.9.9
numpy<2
scenedetect[opencv]==0.6.4
open-clip-torch==2.24.0
Pillow==10.2.0
openai-whisper==20231117
insightface==0.7.3
onnxruntime==1.16.3

# Testing
pytest==8.0.0
pytest-timeout==2.2.0
